doc: clarify each function's purpose in docstring
Is a given function a transducer, a reducer, or a source? Now users can
see at-a-glance both in in-editor help windows and in the published docs
what kind of function each one is.

This is good since there isn't otherwise any separate namespacing to say
"these functions are transducers and these are reducers", etc.